+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 v0.6.1 (2017-11-08)
 ...................
 * prevent duplicate validators, #101
+* add ``always`` kwarg to validators, #102
 
 v0.6.0 (2017-11-07)
 ...................